add findCourseByName,Login auth

+++ backend/src/main/java/com/example/demo/model/dto/CategoryDTO.java
@@ -15,6 +15,6 @@ public class CategoryDTO {
 
     private String description;
 
-    private Integer courseCount;
+    private Long courseCount;
 
 }
